fix bug - Long -> Integer in CategoryRepository declaration

+++ src/main/java/pl/jdomanski/k47/category/CategoryRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
     
     public Category findByName(String name);
     public List<Category> findByType(CategoryType type);
